Add  M java/darts/.gradle/8.7/executionHistory/executionHistory.bin  M java/darts/.gradle/8.7/executionHistory/executionHistory.lock  M java/darts/.gradle/8.7/fileHashes/fileHashes.bin  M java/darts/.gradle/8.7/fileHashes/fileHashes.lock  M java/darts/.gradle/8.7/fileHashes/resourceHashesCache.bin  M java/darts/.gradle/buildOutputCleanup/buildOutputCleanup.lock  M java/darts/build/classes/java/main/Darts.class  M java/darts/build/classes/java/test/DartsTest.class  M java/darts/build/reports/tests/test/classes/DartsTest.html  M java/darts/build/reports/tests/test/index.html  M java/darts/build/reports/tests/test/packages/default-package.html  M java/darts/build/test-results/test/TEST-DartsTest.xml  M java/darts/build/test-results/test/binary/results.bin  M java/darts/build/tmp/compileJava/previous-compilation-data.bin  M java/darts/build/tmp/compileTestJava/previous-compilation-data.bin  M java/darts/src/main/java/Darts.java  M java/darts/src/test/java/DartsTest.java ?? java/darts/build/tmp/compileJava/compileTransaction/ ?? java/darts/build/tmp/compileTestJava/compileTransaction/

+++ java/darts/src/main/java/Darts.java
@@ -1,6 +1,6 @@
 class Darts {
     int score(double xOfDart, double yOfDart) {
-      int distance = (int) Math.sqrt(xOfDart * xOfDart + yOfDart * yOfDart);
+      double distance = Math.sqrt(xOfDart * xOfDart + yOfDart * yOfDart);
       return distance > 10 ? 0 : distance > 5 ? 1 : distance > 1 ? 5 : 10;
     }
 }
